+++ src/app/app.module.ts
@@ -19,6 +19,8 @@ import {
 } from '@angular/material';
 
 
+import {MatTooltipModule} from '@angular/material/tooltip'; 
+
 import { AngularSplitModule } from 'angular-split';
 
 import { AppComponent } from './app.component';
@@ -59,6 +61,7 @@ import { WidgetTutorialComponent } from './widget-tutorial/widget-tutorial.compo
 import { ResetConfigComponent } from './reset-config/reset-config.component';
 import { WidgetStateComponent } from './widget-state/widget-state.component';
 import { ModalWidgetComponent } from './modal-widget/modal-widget.component';
+import { WidgetSwitchComponent } from './widget-switch/widget-switch.component'
 import { ModalPathSelectorComponent } from './modal-path-selector/modal-path-selector.component';
 import { ModalUnitSelectorComponent } from './modal-unit-selector/modal-unit-selector.component';
 import { ObjectKeysPipe } from './object-keys.pipe'
@@ -100,6 +103,7 @@ const appRoutes: Routes = [
     ResetConfigComponent,
     WidgetStateComponent,
     ModalWidgetComponent,
+    WidgetSwitchComponent,
     ModalPathSelectorComponent,
     ModalUnitSelectorComponent,
     ObjectKeysPipe
@@ -115,6 +119,7 @@ const appRoutes: Routes = [
 
     MatMenuModule,
     MatButtonModule,
+    MatTooltipModule,
     MatDialogModule,
     MatSelectModule,
     MatToolbarModule,
@@ -133,6 +138,7 @@ const appRoutes: Routes = [
     WidgetWindComponent,
     WidgetGaugeComponent,
     WidgetStateComponent,
+    WidgetSwitchComponent,
     WidgetTutorialComponent,
     
     //dialogs