+++ src/app/signalk-connection.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { Subscription ,  Observable ,  Subject ,  BehaviorSubject } from 'rxjs';
-import { HttpClient, HttpErrorResponse } from '@angular/common/http';
+import { HttpClient, HttpErrorResponse, HttpHeaders } from '@angular/common/http';
 
 import { AppSettingsService } from './app-settings.service';
 import { SignalKService } from './signalk.service';
@@ -46,13 +46,15 @@ export interface signalKStatus {
 @Injectable()
 export class SignalKConnectionService {
 
+
     // Main URL Variables
     signalKURL: string;
+    signalKToken: string;
     endpointREST: string;
     endpointWS: string;
 
     // Websocket
-    webSocket: WebSocket;
+    webSocket: WebSocket = null;
 
     // status
     signalKStatus: BehaviorSubject<signalKStatus> = new BehaviorSubject<signalKStatus>({
@@ -93,10 +95,22 @@ export class SignalKConnectionService {
     {
         // when signalKUrl changes, do stuff
         this.AppSettingsService.getSignalKURLAsO().subscribe(
-            newURL => {
-                this.signalKURL = newURL;
-                this.resetSignalK();
+          newURL => {
+            this.signalKURL = newURL;
+            if (this.webSocket !== null) {
+              
             }
+            this.resetSignalK();
+          }
+        );
+        // when token changes, do stuff
+        this.AppSettingsService.getSignalKTokenAsO().subscribe(
+          newToken => {
+            this.signalKToken = newToken;
+            if (this.webSocket !== null) {
+              this.webSocket.close();
+            }
+          }
         );
     }
 
@@ -184,8 +198,13 @@ export class SignalKConnectionService {
             setTimeout(()=>{ this.connectEndpointWS();}, 3000);
             return;
         }
+
+        let endpointArgs = "?subscribe=all";
+        if ((this.signalKToken !== null)&&(this.signalKToken != "")) {
+          endpointArgs += "&token="+this.signalKToken;
+        }
         this.webSocketStatusMessage.next("Connecting...");
-        this.webSocket = new WebSocket(this.endpointWS+"?subscribe=all");
+        this.webSocket = new WebSocket(this.endpointWS+endpointArgs);
         this.webSocket.onopen = function (event){
             this.webSocketStatusOK.next(true);
             this.webSocketStatusMessage.next("Connected");
@@ -209,6 +228,15 @@ export class SignalKConnectionService {
     }
 
 
+    publishDelta(message: string) {
+
+      if (!this.webSocketStatusOK.value) {
+        console.log("Tried to publish delta while not connected to Websocket");
+        return;
+      }
+      this.webSocket.send(message);
+    }
+
     //borring stuff, return observables etc
 
     getEndpointAPIStatus() {
@@ -229,5 +257,6 @@ export class SignalKConnectionService {
     getEndpointRESTMessage() {
         return this.restStatusMessage.asObservable();
     }
+
 
 }