+++ src/app/widget-state/widget-state.component.ts
@@ -4,6 +4,7 @@ import {MatDialog,MatDialogRef,MAT_DIALOG_DATA } from '@angular/material';
 
 import { ModalWidgetComponent } from '../modal-widget/modal-widget.component';
 import { SignalKService } from '../signalk.service';
+import { SignalkRequestsService } from '../signalk-requests.service';
 import { WidgetManagerService, IWidget, IWidgetConfig } from '../widget-manager.service';
 
 
@@ -17,14 +18,18 @@ const defaultConfig: IWidgetConfig = {
       pathType: "boolean",
     }
   },
-  selfPaths: true
+  selfPaths: true,
+  putEnable: false,
+  putMomentary: false,
+  putMomentaryValue: true
+  
 };
 
 
 @Component({
   selector: 'app-widget-state',
   templateUrl: './widget-state.component.html',
-  styleUrls: ['./widget-state.component.css']
+  styleUrls: ['./widget-state.component.css', './widget-state.component.scss']
 })
 export class WidgetStateComponent implements OnInit, OnDestroy {
 
@@ -36,11 +41,13 @@ export class WidgetStateComponent implements OnInit, OnDestroy {
   config: IWidgetConfig;
 
   state: boolean = null;
-
+  pressed = false;
+  timeoutHandler;
 
   constructor(
     public dialog:MatDialog,
     private SignalKService: SignalKService,
+    private SignalkRequestsService: SignalkRequestsService,
     private WidgetManagerService: WidgetManagerService) {
   }
 
@@ -72,8 +79,6 @@ export class WidgetStateComponent implements OnInit, OnDestroy {
     );
   }
 
-
-
   unsubscribePath() {
     if (this.valueSub !== null) {
       this.valueSub.unsubscribe();
@@ -82,7 +87,7 @@ export class WidgetStateComponent implements OnInit, OnDestroy {
     }
   }
 
-  openWidgetSettings(content) {
+  openWidgetSettings() {
       
     let dialogRef = this.dialog.open(ModalWidgetComponent, {
       width: '80%',
@@ -101,6 +106,68 @@ export class WidgetStateComponent implements OnInit, OnDestroy {
 
     });
   }
+
+
+
+  handleClickDown() {
+    if (!this.config.putEnable) { return; }
+
+    if (!this.config.putMomentary) { 
+      //on/off mode. Send whatever we're not :)
+      this.SignalkRequestsService.putRequest(
+        this.config.paths['boolPath'].path, 
+        this.config.paths['boolPath'].source, 
+        !this.state);
+        return;
+    } else {
+      // momentary mode
+      this.pressed = true;
+
+      // send it once to start
+      this.SignalkRequestsService.putRequest(
+        this.config.paths['boolPath'].path, 
+        this.config.paths['boolPath'].source, 
+        this.config.putMomentaryValue);
+      //send it again every 20ms
+      this.timeoutHandler = setInterval(() => {
+        this.SignalkRequestsService.putRequest(
+          this.config.paths['boolPath'].path, 
+          this.config.paths['boolPath'].source, 
+          this.config.putMomentaryValue);
+        }, 100);
+
+      return;
+    }
+    
+  }
+
+
+  handleClickUp() {
+    if (!this.config.putEnable || !this.pressed) { return; }
+
+    if (this.config.putMomentary) { 
+      this.pressed = false;
+      clearInterval(this.timeoutHandler);
+      // momentary mode
+      this.SignalkRequestsService.putRequest(
+        this.config.paths['boolPath'].path, 
+        this.config.paths['boolPath'].source, 
+        !this.config.putMomentaryValue);
+        return;
+    }
+    
+  }
+
+
+
+
+
+
+
+
+
+
+
 
 
 }