+++ CHANGELOG.md
@@ -1,8 +1,10 @@
-
-
 ### V0.1.9
 * Updated NPM dependencies
 * Show connection lost overlay on closed websocket connection
+* Only resize Steelseries Gauges max 1 a second (helps in resizing)
+* Add Frequency Units
+* Put Support! On/Off and Momentary support on boolean state
+* Request R/W token from SignalK Server
 
 ### V0.1.8
 * Fix bug related to source parsing in deltas